Extract PublicHolidayService Interface
